Content length is managed internally as a long. Fix a few places that were unnecessarily restricting it to an int.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/tc7.0.x/trunk@1518197 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/coyote/Request.java
@@ -272,7 +272,7 @@ public final class Request {
     }
 
 
-    public void setContentLength(int len) {
+    public void setContentLength(long len) {
         this.contentLength = len;
     }
